+++ tests/unit/Codeception/Module/TestsForWeb.php
@@ -760,4 +760,16 @@ abstract class TestsForWeb extends \PHPUnit_Framework_TestCase
         $this->assertEquals('value2', $form['button2']);
         $this->assertEquals('bob', $form['username']);
     }
+
+    /*
+     * https://github.com/Codeception/Codeception/issues/1274
+     */
+    public function testSubmitFormWithDocRelativePathForAction()
+    {
+        $this->module->amOnPage('/form/example12');
+        $this->module->submitForm('form', array(
+            'test' => 'value'
+        ));
+        $this->module->seeCurrentUrlEquals('/form/example11');
+    }
 }